CLN : Replace array editor with literalinclude in factories basic
	new file:   docs/source/traitsui_user_manual/examples/array_editor.py
	modified:   docs/source/traitsui_user_manual/factories_basic.rst
	deleted:    examples/tutorials/doc_examples/examples/array_editor.py
